Update `unist-util-is`
Breaking: this removes the identity check.  Previously, it was possible
to pass a node in, and remove that node, as the comparison was done
with strict equality (`===`).  `unist-util-is@2.0.0` no longer supports
this and treats a given object (or node) as a shallow partial property
check.

If your code relies on that, you must change it along the following
lines:

```diff
-remove(tree, thing);
+remove(tree, node => node === thing)
```

Closes GH-1.